refactor: AdminViewController와 UserViewController의 패키지 경로를 roomescape.view로 변경
