+++ src/main/java/roomescape/view/AdminViewController.java
@@ -1,4 +1,4 @@
-package roomescape.admin;
+package roomescape.view;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
