+++ src/main/java/roomescape/view/UserViewController.java
@@ -1,4 +1,4 @@
-package roomescape.user;
+package roomescape.view;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
